- Fix emptydisk size (should be string) - Share code for at least some uses of subprocess.run Correctly get run_host.
