+++ clusterbuster
@@ -1299,6 +1299,8 @@ function get_run_started() {
 }
 
 function set_run_aborted() {
+    echo "Run aborted: ${*:-unknown reason}" 1>&2
+    stack_trace
     touch "${cb_tempdir}/___run_aborted"
 }
 
@@ -1312,14 +1314,16 @@ function get_run_failed() {
 
 function set_run_failed() {
     if get_run_failed ; then
-	echo "Secondary failure: ${*:-unknown reason}" 1>&2
-	stack_trace 1 2 1>&2
+	if ! get_run_aborted ; then
+	    echo "Secondary failure: ${*:-unknown reason}" 1>&2
+	    stack_trace 1 2 1>&2
+	fi
 	return
     fi
-    echo "Setting failure status: ${*:-unknown reason}" 1>&2
+    echo "Run failed: ${*:-unknown reason}" 1>&2
     stack_trace
     touch "${cb_tempdir}/___run_failed"
-    run_on_sync_n sh -c "echo 'Please see logs for details.' >> '$sync_error_file'" || set_run_aborted
+    run_on_sync_n sh -c "echo 'Please see logs for details.' >> '$sync_error_file'" || touch "${cb_tempdir}/___run_aborted"
 }
 
 function supports_metrics() {
@@ -1714,7 +1718,7 @@ function _log_helper() {
 	# Error from server: error dialing backend: write tcp 192.168.216.10:55618->192.168.216.13:10250: use of closed network connection
 	# which will result in oc/kubectl get pod (or much of anything else) returning
 	# effectively an empty status.  So we want to make sure that our call to get
-	# pod status only comes after a successful 
+	# pod status only comes after a successful
 	until (( "$(____OC get pod -A --no-headers 2>/dev/null | wc -l)" > 10)) ; do
 	    get_run_failed && return 1
 	    sleep 5
@@ -2794,9 +2798,9 @@ EOF
 function _really_create_objects() {
     if [[ $doit -ne 0 && -n $accumulateddata ]] ; then
 	__OC --validate=false apply -f - 2>&1 <<< "$accumulateddata" || {
-	    echo "Create objects failed:" 2>&1
-	    echo "$accumulateddata" 2>&1
-	    set_run_failed "Failed to create objects"
+	    echo "Create objects failed:" 1>&2
+	    echo "$accumulateddata" 1>&2
+	    set_run_aborted "Failed to create objects"
 	    return 1
 	}
     fi
@@ -3266,7 +3270,7 @@ EOF
 		cat << EOF
 - name: emptydisk-${id}
   emptyDisk:
-    capacity: ${vm_emptydisks[$mountpoint]}
+    capacity: "${vm_emptydisks[$mountpoint]}"
 EOF
 	    done
 	fi
@@ -3965,7 +3969,10 @@ function do_logging() {
 	local -i itimeout=$timeout
 	local -i finis=0
 	while (( timeout < 0 || timeout-- )) ; do
-	    get_run_failed && break
+	    if get_run_failed ; then
+		finis=-1
+		break
+	    fi
 	    # CreateContainerError is not necessarily fatal
 	    if __OC get pods -A -l "$basename" |grep -q -E -e '([^r]Error|Evicted|Crash)' ; then
 		echo "Run failed:" 1>&2
@@ -4101,7 +4108,7 @@ function run_clusterbuster_2() {
     set_run_started
 
     do_logging || return 1
-    if ((! status)) ; then
+    if ! get_run_failed || ((! status)) ; then
 	report_object_creation
     fi
     return 0
@@ -4127,6 +4134,7 @@ function run_clusterbuster_1() {
 	    retrieve_prometheus_snapshot "$artifactdir"
 	fi
     fi
+    get_run_failed && status=1
     return $status
 }
